+++ src/prisma/generator/templates/client.py.jinja
@@ -355,15 +355,19 @@ class {{ model.name }}BatchActions:
     def create_many(
         self,
         data: List[types.{{ model.name }}CreateWithoutRelationsInput],
+        {% if active_provider not in ['mongodb', 'sqlserver', 'sqlite'] %}
         *,
         skip_duplicates: Optional[bool] = None,
+        {% endif %}
     ) -> None:
         self._batcher._add(
             method='create_many',
             model=models.{{ model.name }},
             arguments={
                 'data': data,
+                {% if active_provider not in ['mongodb', 'sqlserver', 'sqlite'] %}
                 'skipDuplicates': skip_duplicates,
+                {% endif %}
             },
             root_selection=['count'],
         )
